requirements.txt
streamlit
panda
numpy
librosa
soundfile
plotly
spotdl
xlsxwriter